All tasks are done
Created the CityGrid class, developed an algorithm for placing towers and functions for visualizing CityGrid, solved the optimization problem. All variable parameters in the code can be entered manually (for this you need auto=False) or selected randomly automatically (configured initially). The parameter of the percentage coverage of the city can be changed